+++ VILLAOS1/commands/command1.asm
@@ -1,0 +1,3 @@
+
+
+Unknown token '﻿' found at 1/0.
